upgrade dependencies and fix transparency handling
Upgrade SDK version and dart image library version for
better compatibility with other plugins.

Before this change the alpha channel was ignored when
calculating the qr code bit matrix. A pixel with color
(r: 0, g: 0, b: 0, a: 0) was intepreted as a black pixel (set bit)
although its transparent and therefore not visible (unset bit)
when rendered. Now the image is first converted from rgba to rgb
by blending it on a white background.

Add convenience method to QrCode for scanning an encoded image
in any format supported by the image package.